Better naming convention for hacky solution
