--- home/homePane.js
@@ -1,1 +1,0 @@
-module.exports = require('./homePane.ts').default
